fix: item create/update always failed validation regardless of input
Found while testing contact/warehouse/category/item creation end to end.
CreateItemCommand and UpdateItemCommand declared Type as plain int instead of
the ItemType enum (every other command in the codebase, e.g.
CreateInvoiceCommand.Type, correctly uses the real enum type per
PROJECT_RULES.md's own "use native enum types in Command/Query DTOs" rule —
Item was the outlier). Both validators call `.IsInEnum()` on that field,
which only means something against an actual enum type; run against a plain
int it always evaluated to invalid, so POST/PUT /api/items rejected every
request no matter what Type value was sent. The 19 seeded items never went
through this endpoint (DataSeeder inserts them directly), which is why this
went unnoticed.

Changed Type to ItemType in both commands, dropped the now-redundant
(ItemType) casts in the handlers, and simplified the validators'
`x.Type == (int)ItemType.FixedAsset` checks to compare enums directly.

Added Integration/ItemEndpointsIntegrationTests — a real HTTP POST through
the full validator pipeline, which is the only kind of test that could have
caught this (existing unit tests call CreateItemHandler directly, bypassing
FluentValidation entirely).

Verified: dotnet build (0 errors), dotnet test (84/84), and a live
docker-compose POST /api/items call returning 201.

+++ backend/Accounting.Application/Items/Commands/Create/CreateItemValidator.cs
@@ -36,7 +36,7 @@ public class CreateItemValidator : AbstractValidator<CreateItemCommand>
         RuleFor(x => x.UsefulLifeYears)
             .GreaterThan(0)
             .LessThanOrEqualTo(50)
-            .When(x => x.Type == (int)ItemType.FixedAsset && x.UsefulLifeYears.HasValue)
+            .When(x => x.Type == ItemType.FixedAsset && x.UsefulLifeYears.HasValue)
             .WithMessage("Demirbaşlar için faydalı ömür 1-50 yıl arası olmalıdır");
     }
 }
